+++ runtime/src/main/kotlin/com/roomql/runtime/Column.kt
@@ -1,3 +1,3 @@
 package com.roomql.runtime
 
-public data class Column<T>(val columnName: String, val tableName: String)
+public data class Column<T>(val columnName: String, val tableName: String) : Expression<T>